+++ src/REvoDesign/driver/ui_driver.py
@@ -9,6 +9,7 @@ import shutil
 from collections.abc import Callable
 from dataclasses import dataclass
 from functools import partial, wraps
+import time
 from typing import Any, Protocol, TypeVar, overload
 
 import omegaconf.errors
@@ -211,13 +212,43 @@ class Config:
         """
         save_configuration(self.cfg, self.name)
 
+    def save_to_experiment(self, experiment: str):
+        '''
+        located at REVODESIGN_CONFIG_DIR
+
+        experiment:
+            - experiment: experiment/<name-of-experiment>.yaml
+            - cache: cache/<name-of-cache>.yaml
+        return:
+            - path to the saved experiment file
+        '''
+        save_configuration(self.cfg, experiment)
+        return os.path.join(REVODESIGN_CONFIG_DIR, f'{experiment}.yaml')
+
     def reload(self):
         """
         Reloads the configuration data from the configuration file.
         """
         self.cfg = reload_config_file(self.name)
 
-    def reload_from(self, path: str):
+    def reload_from_experiment(self, experiment: str):
+        '''
+        experiment:
+            - experiment: experiment/<name-of-experiment>.yaml
+            - cache: cache/<name-of-cache>.yaml
+        '''
+        reload_from_prefix=experiment.split('/',1)[0]
+        if reload_from_prefix not in ('experiments', 'cache'):
+            raise issues.UnexpectedWorkflowError(f'The reload type must be experiments or cache, not {reload_from_prefix}')
+        
+        # follow the hydra rule
+        expected=os.path.join(REVODESIGN_CONFIG_DIR, f'{experiment}.yaml')
+        if not os.path.isfile(expected):
+            raise FileNotFoundError(f'An experiment file is expected at {expected} for {experiment}')
+        self.cfg=reload_config_file(config_name=experiment)[reload_from_prefix]
+
+
+    def reload_from_path(self, path: str):
         """
         Reloads the configuration data from a specified file path.
         Args:
@@ -246,13 +277,16 @@ class Config:
             file_path (str): The path to save the configuration file.
 
         """
+        # get a uniq timestamp
+        timestamp = time.strftime("%Y%m%d_%H%M%S", time.localtime())
+        
+        cached_path=self.save_to_experiment(f'cache/{self.name}_cached_{timestamp}')
 
-        # save to disk first
-        self.save()
+        # move to the target path
+        # unless explicitly saved, the self.path wont be saved to
 
-        # copy to the target path
-        shutil.copy(self.path, file_path)
-        logging.info(f"Config file {self.name} saved to {file_path}")
+        shutil.move(cached_path, file_path)
+        logging.info(f"Config file {self.name} saved to {file_path} (cached at {cached_path})")
 
 
 def require_non_headless(method):