+++ tests/bootstrap/test_set_config.py
@@ -3,6 +3,10 @@ from unittest.mock import MagicMock, patch
 import pytest
 from omegaconf import DictConfig, OmegaConf
 
+# from tests.conftest import check_real_config_dir
+
+# check_real_config_dir() # failed
+
 from REvoDesign.bootstrap.set_config import (
     ConfigConverter,
     experiment_config,
@@ -11,15 +15,20 @@ from REvoDesign.bootstrap.set_config import (
     save_configuration,
     set_cache_dir,
     set_REvoDesign_config_file,
+
 )
 
 
+
 def test_set_REvoDesign_config_file():
+    # check_real_config_dir() # failed
     with (
-        patch("REvoDesign.bootstrap.set_config.os.path.exists", return_value=False),
-        patch("REvoDesign.bootstrap.set_config.glob.glob", return_value=[]),
+        patch("REvoDesign.bootstrap.set_config.os.path.isfile", return_value=False),
+        patch("REvoDesign.bootstrap.set_config.os.path.isdir", return_value=False),
+        patch("REvoDesign.bootstrap.set_config.os.listdir", return_value=[]),
         patch("REvoDesign.bootstrap.set_config.shutil.copytree") as mock_copytree,
     ):
+        
 
         main_config = set_REvoDesign_config_file()
         mock_copytree.assert_called_once()
@@ -84,3 +93,37 @@ def test_is_package_installed():
 
     with patch("REvoDesign.bootstrap.set_config.importlib.util.find_spec", return_value=None):
         assert is_package_installed("nonexistent") is False
+
+@pytest.mark.parametrize('user_decide',[
+    True, False
+])
+
+def test_main_config_upgrade(user_decide, patch_config_user_data):
+    with (
+        patch('REvoDesign.bootstrap.set_config.decide', return_value=user_decide) as mock_decide,
+        
+        # main.yaml not exists
+        patch("REvoDesign.bootstrap.set_config.os.path.isfile", return_value=False),
+        
+        # configure dir exists
+        patch("REvoDesign.bootstrap.set_config.os.path.isdir", return_value=True),
+
+        # configure dir exists
+        patch("REvoDesign.bootstrap.set_config.os.listdir", return_value=[ 'master.yaml', 'a_dir']),
+
+        # if user has decide, the tree will be removed
+        patch("REvoDesign.bootstrap.set_config.shutil.rmtree") as mock_rmtree,
+
+        # whatever, the new tree will be copied
+        patch("REvoDesign.bootstrap.set_config.shutil.copytree") as mock_copytree,
+        ):
+
+            main_config = set_REvoDesign_config_file()
+            mock_decide.assert_called_once()
+            mock_copytree.assert_called_once()
+            if user_decide:
+                mock_rmtree.assert_called_once()
+            else:
+                mock_rmtree.assert_not_called()
+
+